chore(virtctl): Move fs list command to own file
