+++ pkg/virtctl/vm/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "add_volume.go",
         "expand.go",
+        "fs_list.go",
         "migrate.go",
         "migrate_cancel.go",
         "remove_volume.go",
